+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 ## [0.32.0] - xxx
+### Added
+- Implemented ```date-format```
 
 ## [0.31.0] - 2021-08-22 
 ### Added